# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/customers": {
      "get": {
        "operationId": "CustomersController_listCustomers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page number, starts from 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size, max 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Fuzzy search keyword",
            "schema": {
              "example": "demo",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "active",
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "blocked"
              ]
            }
          },
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "group_01",
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "tag_01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List customers",
        "tags": [
          "admin-customers"
        ]
      },
      "post": {
        "operationId": "CustomersController_createCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create customer",
        "tags": [
          "admin-customers"
        ]
      }
    },
    "/api/v1/admin/customers/{id}": {
      "get": {
        "operationId": "CustomersController_getCustomer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get customer detail",
        "tags": [
          "admin-customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_updateCustomer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update customer profile",
        "tags": [
          "admin-customers"
        ]
      }
    },
    "/api/v1/admin/customers/{id}/status": {
      "patch": {
        "operationId": "CustomersController_updateCustomerStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update customer status",
        "tags": [
          "admin-customers"
        ]
      }
    },
    "/api/v1/admin/customers/{id}/reset-password": {
      "post": {
        "operationId": "CustomersController_resetCustomerPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetCustomerPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reset customer password",
        "tags": [
          "admin-customers"
        ]
      }
    },
    "/api/v1/admin/customer-groups": {
      "get": {
        "operationId": "CustomersController_listGroups",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page number, starts from 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size, max 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Fuzzy search keyword",
            "schema": {
              "example": "demo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerGroupListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List customer groups",
        "tags": [
          "admin-customers"
        ]
      },
      "post": {
        "operationId": "CustomersController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerGroupSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create customer group",
        "tags": [
          "admin-customers"
        ]
      }
    },
    "/api/v1/admin/customer-groups/{id}": {
      "patch": {
        "operationId": "CustomersController_updateGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerGroupSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update customer group",
        "tags": [
          "admin-customers"
        ]
      }
    },
    "/api/v1/admin/customer-tags": {
      "get": {
        "operationId": "CustomersController_listTags",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page number, starts from 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size, max 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Fuzzy search keyword",
            "schema": {
              "example": "demo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerTagListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List customer tags",
        "tags": [
          "admin-customers"
        ]
      },
      "post": {
        "operationId": "CustomersController_createTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerTagSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create customer tag",
        "tags": [
          "admin-customers"
        ]
      }
    },
    "/api/v1/admin/customer-tags/{id}": {
      "patch": {
        "operationId": "CustomersController_updateTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerTagSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update customer tag",
        "tags": [
          "admin-customers"
        ]
      }
    },
...
  },
  "info": {
    "title": "rtnn backend api",
    "description": "Template backend kernel for RTNN",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 57
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "LabeledReferenceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "grp_01JABCD123"
          },
          "name": {
            "type": "string",
            "example": "VIP Customers"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "CustomerSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cus_01JABCD123"
          },
          "accountId": {
            "type": "string",
            "example": "acc_01JABCD123"
          },
          "email": {
            "type": "string",
            "example": "customer@example.com"
          },
          "name": {
            "type": "string",
            "example": "Customer Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "blocked"
            ],
            "example": "active"
          },
          "tenantId": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "phone": {
            "type": "object",
            "example": "+1-555-0100",
            "nullable": true
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabeledReferenceDto"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabeledReferenceDto"
            }
          },
          "groupNames": {
            "example": [
              "VIP Customers"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tagNames": {
            "example": [
              "High value"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastLoginAt": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "accountId",
          "email",
          "name",
          "status",
          "tenantId",
          "groups",
          "tags",
          "lastLoginAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "CustomerListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CustomerDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cus_01JABCD123"
          },
          "accountId": {
            "type": "string",
            "example": "acc_01JABCD123"
          },
          "email": {
            "type": "string",
            "example": "customer@example.com"
          },
          "name": {
            "type": "string",
            "example": "Customer Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "blocked"
            ],
            "example": "active"
          },
          "tenantId": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "phone": {
            "type": "object",
            "example": "+1-555-0100",
            "nullable": true
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabeledReferenceDto"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabeledReferenceDto"
            }
          },
          "groupNames": {
            "example": [
              "VIP Customers"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tagNames": {
            "example": [
              "High value"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastLoginAt": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "notes": {
            "type": "object",
            "example": null,
            "nullable": true
          }
        },
        "required": [
          "id",
          "accountId",
          "email",
          "name",
          "status",
          "tenantId",
          "groups",
          "tags",
          "lastLoginAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@example.com"
          },
          "name": {
            "type": "string",
            "example": "Template Customer"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "StrongPass123!"
          },
          "phone": {
            "type": "string",
            "example": "13800138000"
          },
          "tenantId": {
            "type": "string",
            "example": "default"
          },
          "groupIds": {
            "example": [
              "group_01"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tagIds": {
            "example": [
              "tag_01"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@example.com"
          },
          "name": {
            "type": "string",
            "example": "Template Customer"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "StrongPass123!"
          },
          "phone": {
            "type": "string",
            "example": "13800138000"
          },
          "tenantId": {
            "type": "string",
            "example": "default"
          },
          "groupIds": {
            "example": [
              "group_01"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tagIds": {
            "example": [
              "tag_01"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdateCustomerStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "blocked"
            ],
            "example": "inactive"
          }
        },
        "required": [
          "status"
        ]
      },
      "ResetCustomerPasswordDto": {
        "type": "object",
        "properties": {
          "nextPassword": {
            "type": "string",
            "minLength": 8,
            "example": "StrongPass123!"
          }
        },
        "required": [
          "nextPassword"
        ]
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
      "CustomerGroupSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "grp_01JABCD123"
          },
          "name": {
            "type": "string",
            "example": "VIP Customers"
          },
          "description": {
            "type": "string",
            "example": "Customers with higher support priority"
          },
          "memberCount": {
            "type": "number",
            "example": 3
          },
          "customerCount": {
            "type": "number",
            "example": 3
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "CustomerGroupListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerGroupSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateCustomerGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "VIP Customers"
          },
          "slug": {
            "type": "string",
            "example": "vip-customers"
          },
          "description": {
            "type": "string",
            "example": "High value customers"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCustomerGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "VIP Customers"
          },
          "slug": {
            "type": "string",
            "example": "vip-customers"
          },
          "description": {
            "type": "string",
            "example": "High value customers"
          }
        }
      },
      "CustomerTagSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "tag_01JABCD123"
          },
          "name": {
            "type": "string",
            "example": "High value"
          },
          "color": {
            "type": "object",
            "example": "#0ea5e9",
            "nullable": true
          },
          "usageCount": {
            "type": "number",
            "example": 5
          },
          "customerCount": {
            "type": "number",
            "example": 5
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "CustomerTagListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerTagSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateCustomerTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Potential"
          },
          "slug": {
            "type": "string",
            "example": "potential"
          },
          "color": {
            "type": "string",
            "example": "#60a5fa"
          },
          "description": {
            "type": "string",
            "example": "Potential customer segment"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCustomerTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Potential"
          },
          "slug": {
            "type": "string",
            "example": "potential"
          },
          "color": {
            "type": "string",
            "example": "#60a5fa"
          },
          "description": {
            "type": "string",
            "example": "Potential customer segment"
          }
        }
      },
...
    }
  }
}